AgentForge/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentforge"
version = "0.1.0"
description = "A modular agent development framework for building and assembling agent components"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "AgentForge Team" }
]
keywords = ["agent", "ai", "llm", "framework", "tool-use"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
mcp = [
    "mcp>=0.1.0",
]

[project.scripts]
agentforge = "core.channel.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/core"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
